Apply class extension within HandleCopyTermDefaultsToDefaultObject (#39)
We have a custom FieldNotify implementation that relies on fetching the CDO of its owning widget class in order to set up bindings based on its serialized properties. When compiling its associated widget blueprint however, it turns out that HandleFinishCompilingClass triggers before the CDO has had a chance to fully serialize. This causes its associated bindings to get orphaned every time the blueprint is loaded or recompiled.

Switching this over to HandleCopyTermDefaultsToDefaultObject resolves this issue gracefully, as it triggers at a later point during blueprint compilation.