Updated UMDFastBindingDestination_Property::SetupBindingItems to ensure the PathRoot node has a valid ItemProperty
There was an issue present where `CheckCachedNeedsUpdate()` within `UMDFastBindingDestination_Property::UpdateDestination_Internal` would succeed every time, leading to the value getting reapplied each tick (and consequently broadcasting the field notify every time).

Since `SetupBindingItems` passed in nullptr as the `ItemProperty`, `FMDFastBindingItem::GetValue` would hit the `ItemProp == nullptr` branch every time. This prevented it from being able to allocate a default value and thus only update when it was needed.

It's worth noting that `UMDFastBindingDestination_Property` supports both objects and structs for its path root, so this change will introduce a minor UX regression in the Binding Editor, in that the Path Root node will be displayed as blue (for UObject) instead of grey (for wildcard).